Implement Flask with deposit flow
- Introduced Flask to create a web app of the game in _main.py_.
- Renamed old _main.py_ tp _slot-machine-game.py_. (contains game logic)

- Created working home page with deposit form.
- Flask app asks the user to enter an initial deposit amount.
- Web app displays user confirmation message after deposit.
- App validates deposit input and stored balance in session.
- Created a "Start Playing" button to route user to /play endpoint.
- Created placeholder /play route to prepare for gameplay screen.
